(feat:nick) AccountService controller refinement

+++ src/main/java/service/AccountService.java
@@ -21,17 +21,25 @@ public class AccountService {
       * exists with the 'Account' s respective fields
     */
 
+    public Boolean checkDatabaseForExistingUsername(String userInput) {
+        return readByUsername(userInput) != null; //There's an account with that field
+    }
+
+    public Boolean checkDatabaseForExistingEmail(String userInput) {
+        return readByEmail(userInput) != null; //There's an account with that field
+    }
+
     public Account create(Account account) {
+        if (checkDatabaseForExistingEmail(account.getAppEmail()) ||
+                checkDatabaseForExistingUsername(account.getUsername())) {
+            throw new ResourceNotFoundException("An account already exists with that information");
+        }
         return repo.save(account);
     }
 
     public Account read(Long id) {
-        try {
-            return repo.findById(id).get();
-        } catch (Exception e) {
-            throw new ResourceNotFoundException("There is " +
-                    "no account with that ID");
-        }
+        return repo.findById(id).
+                orElseThrow(() -> new ResourceNotFoundException("User does not exist"));
     }
 
     public Account readByUsername(String username) {
@@ -39,6 +47,20 @@ public class AccountService {
             List<Account> accList = new ArrayList<>();
             readAll().forEach(account -> {
                 if (account.getUsername().equals(username))
+                    accList.add(account);
+            });
+            return accList.get(0);
+        } catch (Exception e) {
+            throw new ResourceNotFoundException("There is no " +
+                    "account with that username");
+        }
+    }
+
+    public Account readByEmail(String userInput) {
+        try {
+            List<Account> accList = new ArrayList<>();
+            readAll().forEach(account -> {
+                if (account.getAppEmail().equals(userInput))
                     accList.add(account);
             });
             return accList.get(0);
@@ -56,20 +78,32 @@ public class AccountService {
     }
 
     public Account updateUsername(Long id, String newUserName) {
-        Account originalAccount = repo.findById(id).get();
+        Account originalAccount = read(id);
         originalAccount.setUsername(newUserName);
         return repo.save(originalAccount);
     }
 
     public Account updateBalance(Long id, double input) { //input can be positive or negative
-        Account originalAccount = repo.findById(id).get();
+        Account originalAccount = read(id);
         double newBalanceOfAcc = originalAccount.getBalance() + input;
         originalAccount.setBalance(newBalanceOfAcc);
         return repo.save(originalAccount);
     }
+
+    public Account updateEmail(Long id, String newEmail) {
+        Account ogAcc = read(id);
+        ogAcc.setUsername(newEmail);
+        repo.save(ogAcc);
+        return ogAcc;
+    }
+
     public Account delete(Account account) {
-        repo.delete(account);
-        return account;
+        if (account.getBalance() > 0.0 || account.getBalance() < 0.0) {
+            throw new ResourceNotFoundException("You must set your balance to $0.00 before deletion");
+        } else {
+            repo.delete(account);
+            return account;
+        }
     }
 
     public Account delete(Long id) {
